Update underactuated.html
I am not sure if you welcome typo corrections in the form of pull requests. This html file might be autogenerated.

I think \succ, \prec should be used for unbound functions (as in V≻0 ), but once you bind it, V(x) evaluates to a scalar, and so I think <,> fit better.